requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==22.2.0
backoff==1.11.1
bcrypt==4.0.1
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
discord-py-slash-command==4.2.1
discord.py==2.2.2
distlib==0.3.7
dnspython==2.4.1
email-validator==2.0.0.post2
filelock==3.12.2
flash==1.0.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Dropzone==1.6.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
heif-image-plugin==0.5.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Mako==1.2.4
MarkupSafe==2.1.2
msgpack==1.0.5
multidict==6.0.4
nextcord==2.5.0
nltk==3.8.1
piexif==1.1.3
Pillow==10.0.0
pillow-heif==0.10.0
platformdirs==3.9.1
protobuf==4.23.3
pycparser==2.21
pygame==2.3.0.dev2
pyheif @ git+https://github.com/carsales/pyheif.git@694c3b2538978047e6c3070a510c03be399bc542
PyNaCl==1.5.0
pynvim==0.4.3
regex==2022.10.31
requests==2.28.2
six==1.16.0
soupsieve==2.4
spotify==0.10.2
SQLAlchemy==2.0.9
tk==0.1.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.24.2
Wavelink==2.4.0
WaveLinkCord==2.2.1
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.8.2
youtube-dl==2021.12.17
